+++ core-auth-fe/src/core-auth/react-admin/LoginPage/index.tsx
@@ -42,8 +42,6 @@ export const csrfToken = () => {
   return meta && meta.content
 }
 
-const csrf = csrfToken()
-
 const AppInitHeadline = () => {
   return (
     <Box sx={{ p: 3 }}>
@@ -209,11 +207,13 @@ export const LoginPage = (props: LoginPageProps) => {
     btnSx,
   }) => {
     return (
-      <Box component="form" method="post" action={href}>
-        <Input type="hidden" name="authenticity_token" value={csrf} />
-        <Button sx={btnSx} type="submit" variant="outlined" startIcon={icon}>
-          {label}
-        </Button>
+      <Box
+        component="form"
+        method='post'
+        action={href}
+      >
+        <Input type='hidden' name='authenticity_token' value={csrfToken()} />
+        <Button sx={btnSx} type='submit' variant="outlined" startIcon={icon}>{label}</Button>
       </Box>
     )
   }