+++ core-webhooks/lib/core/webhooks/controllers/event_controller.rb
@@ -22,10 +22,11 @@ module Core
           end
 
           handler.ensure_authentic!
+          handler.confirm_subscription! if handler.respond_to?(:confirm_subscription!) && handler.respond_to?(:subscribe_event?) && handler.subscribe_event? 
 
           if handler.duplicate?
             # don't persist duplicate events
-            ::Rails.logger.error "Duplicate #{handler.provider} webhook received #{webhook_event.data}."
+            ::Rails.logger.error "Duplicate #{handler.class.provider} webhook received #{webhook_event.data}."
 
             head :ok
             return