+++ core-webhooks/lib/core/webhooks/event.rb
@@ -21,10 +21,10 @@ module Core
         listener_classname = "#{::Core::Webhooks.webhooks_listeners.namespace}/#{webhook_event.source}".camelize
         listener = "::#{listener_classname}".constantize
         listener.send(::Core::Webhooks.webhooks_listeners.process_event_method, webhook_event)
-      rescue NoMethodError => e
-        ::Rails.logger.error "Listener #{listener_classname} does not respond to process_event. Expected it to be defined but it was not."
       rescue NameError => e
-        ::Rails.logger.error "No listener for received #{webhook_event.source} webhook #{webhook_event.event}. Expected #{listener_classname} to be defined but it was not."
+        ::Rails.logger.error "No listener for received #{webhook_event.source} webhook #{webhook_event.event}. Expected #{listener_classname} to be defined but it was not. #{e}"
+      rescue StandardError => e
+        ::Rails.logger.error "Listener #{listener_classname} error processing event #{webhook_event.event}. #{e}"
       end
     end
   end