+++ core-auth-fe/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @groovestack/auth
 
+## 0.2.28
+
+### Patch Changes
+
+- ensure latest csrf token is included in social sign in button posts
+
+## 0.2.27
+
+### Patch Changes
+
+- export social identity icons reference many field
+
 ## 0.2.26
 
 ### Patch Changes