+++ test-core-fe/CHANGELOG.md
@@ -1,5 +1,19 @@
 # test-core-fe
 
+## 0.2.134
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.28
+
+## 0.2.133
+
+### Patch Changes
+
+- Updated dependencies
+  - @groovestack/auth@0.2.27
+
 ## 0.2.132
 
 ### Patch Changes